move handler to manifestSchema and include handlerSchema
