+++ client/src/components/Manifest/Handler/index.ts
@@ -1,8 +1,8 @@
+import { Handler, handlerSchema, Signer } from 'components/Manifest/manifestSchema';
 import { HandlerForm } from './HandlerForm';
 import { HandlerSearchForm } from './HandlerSearchForm';
 import { TransporterSearchForm } from './TransporterSearchForm';
 import { AddHandler } from './AddHandler';
-import { Handler, handlerSchema, Signer } from './handlerSchema';
 
 export { HandlerForm, TransporterSearchForm, HandlerSearchForm, AddHandler, handlerSchema };
 export type { Handler, Signer };
