+++ client/src/components/Manifest/index.ts
@@ -1,5 +1,6 @@
 import { ManifestForm } from './ManifestForm';
 import { Manifest } from './manifestSchema';
+import { Handler } from './manifestSchema';
 
 export { ManifestForm };
-export type { Manifest };
+export type { Manifest, Handler };
