+++ client/src/types/manifest.ts
@@ -1,7 +1,6 @@
-import { RejectionInfo } from 'components/Manifest/manifestSchema';
+import { Handler, RejectionInfo, Signer } from 'components/Manifest/manifestSchema';
 import { WasteLine } from 'components/Manifest/WasteLine/wasteLineSchema';
 import { AdditionalInfo } from 'components/AdditionalInfo';
-import { Handler, Signer } from 'components/Manifest/Handler';
 import { Transporter } from 'components/Manifest/Transporter';
 import { RcraLocality, rcraLocalitySchema } from 'components/RcraSite';
 import { z } from 'zod';
